+++ maXMLian/src/dd/kms/maxmlian/impl/DocumentBuilderFactoryImpl.java
@@ -2,18 +2,17 @@ package dd.kms.maxmlian.impl;
 
 import dd.kms.maxmlian.api.DocumentBuilder;
 import dd.kms.maxmlian.api.DocumentBuilderFactory;
-import dd.kms.maxmlian.api.XmlInputFactoryProvider;
 
 import javax.xml.stream.XMLInputFactory;
 
 public class DocumentBuilderFactoryImpl implements DocumentBuilderFactory
 {
-	private final XMLInputFactory	factory;
+	private final XMLInputFactory	xmlInputFactory;
 	private boolean	reuseInstances	= false;
 
-	public DocumentBuilderFactoryImpl(XmlInputFactoryProvider... xmlInputFactoryProviders) throws IllegalStateException {
-		this.factory = getFirstXmlInputFactory(xmlInputFactoryProviders);
-		factory.setProperty(XMLInputFactory.IS_VALIDATING, false);
+	public DocumentBuilderFactoryImpl(XMLInputFactory xmlInputFactory) throws IllegalStateException {
+		this.xmlInputFactory = xmlInputFactory;
+		xmlInputFactory.setProperty(XMLInputFactory.IS_VALIDATING, false);
 	}
 
 	@Override
@@ -25,7 +24,7 @@ public class DocumentBuilderFactoryImpl implements DocumentBuilderFactory
 	@Override
 	public DocumentBuilderFactory namespaceAware(boolean namespaceAware) throws IllegalStateException {
 		try {
-			factory.setProperty(XMLInputFactory.IS_NAMESPACE_AWARE, namespaceAware);
+			xmlInputFactory.setProperty(XMLInputFactory.IS_NAMESPACE_AWARE, namespaceAware);
 		} catch (IllegalStateException e) {
 			String action = namespaceAware ? "enable" : "disable";
 			throw new IllegalStateException("Cannot " + action + " namespace awareness because the internal StAX parser does not support it.");
@@ -35,26 +34,12 @@ public class DocumentBuilderFactoryImpl implements DocumentBuilderFactory
 
 	@Override
 	public DocumentBuilderFactory normalize(boolean normalize) {
-		factory.setProperty(XMLInputFactory.IS_COALESCING, normalize);
+		xmlInputFactory.setProperty(XMLInputFactory.IS_COALESCING, normalize);
 		return this;
 	}
 
 	@Override
 	public DocumentBuilder newDocumentBuilder() {
-		return new DocumentBuilderImpl(factory, reuseInstances);
-	}
-
-	private static XMLInputFactory getFirstXmlInputFactory(XmlInputFactoryProvider... xmlInputFactoryProviders) throws IllegalStateException {
-		XMLInputFactory factory = null;
-		for (XmlInputFactoryProvider xmlInputFactoryProvider : xmlInputFactoryProviders) {
-			factory = xmlInputFactoryProvider.getXMLInputFactory().orElse(null);
-			if (factory != null) {
-				break;
-			}
-		}
-		if (factory == null) {
-			throw new IllegalStateException("Cannot instantiate an XMLInputFactory");
-		}
-		return factory;
+		return new DocumentBuilderImpl(xmlInputFactory, reuseInstances);
 	}
 }